+++ src/wallet/wallet.h
@@ -988,7 +988,7 @@ public:
                          int &nChangePosInOut, std::string &strFailReason,
                          bool lockUnspents,
                          const std::set<int> &setSubtractFeeFromOutputs,
-                         CCoinControl coinControl, bool keepReserveKey = true);
+                         CCoinControl coinControl, bool keepReserveKey);
     bool SignTransaction(CMutableTransaction &tx);
 
     /**