+++ test/functional/rpc_fundrawtransaction.py
@@ -27,8 +27,10 @@ def get_unspent(listunspent, amount):
 
 class RawTransactionsTest(BitcoinTestFramework):
     def set_test_params(self):
-        self.num_nodes = 4
+        self.num_nodes = 5
         self.setup_clean_chain = True
+        self.extra_args = [[], [], [], [],
+                           ["-deprecatedrpc=fundrawtransaction"]]
 
     def setup_network(self, split=False):
         self.setup_nodes()
@@ -37,6 +39,7 @@ class RawTransactionsTest(BitcoinTestFramework):
         connect_nodes_bi(self.nodes[1], self.nodes[2])
         connect_nodes_bi(self.nodes[0], self.nodes[2])
         connect_nodes_bi(self.nodes[0], self.nodes[3])
+        connect_nodes_bi(self.nodes[0], self.nodes[4])
 
     def run_test(self):
         min_relay_tx_fee = self.nodes[0].getnetworkinfo()['relayfee']
@@ -74,6 +77,7 @@ class RawTransactionsTest(BitcoinTestFramework):
             watchonly_address, watchonly_amount)
         self.nodes[0].sendtoaddress(
             self.nodes[3].getnewaddress(), watchonly_amount / 10)
+        self.nodes[0].sendtoaddress(self.nodes[4].getnewaddress(), 5.0)
 
         self.nodes[0].sendtoaddress(self.nodes[2].getnewaddress(), 1.5)
         self.nodes[0].sendtoaddress(self.nodes[2].getnewaddress(), 1.0)
@@ -476,6 +480,7 @@ class RawTransactionsTest(BitcoinTestFramework):
         self.nodes[1].node_encrypt_wallet("test")
         self.stop_node(2)
         self.stop_node(3)
+        self.stop_node(4)
 
         self.start_nodes()
         # This test is not meant to test fee estimation and we'd like
@@ -487,6 +492,7 @@ class RawTransactionsTest(BitcoinTestFramework):
         connect_nodes_bi(self.nodes[1], self.nodes[2])
         connect_nodes_bi(self.nodes[0], self.nodes[2])
         connect_nodes_bi(self.nodes[0], self.nodes[3])
+        connect_nodes_bi(self.nodes[0], self.nodes[4])
         self.sync_all()
 
         # drain the keypool
@@ -681,20 +687,45 @@ class RawTransactionsTest(BitcoinTestFramework):
             result3['fee'], FromHex(CTransaction(), result3['hex']).billable_size(), 10 * result_fee_rate, 10)
 
         #
+        # DEPRECATED, should be removed in v0.20
         # Test address reuse option #
         #
 
-        result3 = self.nodes[3].fundrawtransaction(
-            rawTx, {"reserveChangeKey": False})
-        res_dec = self.nodes[0].decoderawtransaction(result3["hex"])
+        dInputs = []
+        dOutputs = {self.nodes[4].getnewaddress(): 1}
+        dRawTx = self.nodes[4].createrawtransaction(dInputs, dOutputs)
+        dResult = self.nodes[4].fundrawtransaction(
+            dRawTx, {"reserveChangeKey": False})
+        res_dec = self.nodes[0].decoderawtransaction(dResult["hex"])
         changeaddress = ""
         for out in res_dec['vout']:
             if out['value'] > 1.0:
                 changeaddress += out['scriptPubKey']['addresses'][0]
         assert(changeaddress != "")
-        nextaddr = self.nodes[3].getrawchangeaddress()
+        nextaddr = self.nodes[4].getrawchangeaddress()
         # frt should not have removed the key from the keypool
         assert(changeaddress == nextaddr)
+
+        #
+        # DEPRECATED, should be removed in v0.20
+        # Test address reuse option does #
+        # throws an rpc error when not deprecated #
+        #
+
+        assert_raises_rpc_error(-32, "fundrawtransaction -reserveChangeKey "
+                                + "is deprecated and will be fully removed "
+                                + "in v0.20.  To use the -reserveChangeKey "
+                                + "option in v0.19, restart bitcoind with "
+                                + "-deprecatedrpc=fundrawtransaction.\n"
+                                + "Projects should transition to expecting "
+                                + "change addresses removed from the keypool "
+                                + "before upgrading to v0.20",
+                                self.nodes[3].fundrawtransaction, rawTx,
+                                {"reserveChangeKey": False})
+
+        #
+        # Test no address reuse occurs #
+        #
 
         result3 = self.nodes[3].fundrawtransaction(rawTx)
         res_dec = self.nodes[0].decoderawtransaction(result3["hex"])