+++ doc/release-notes.md
@@ -3,4 +3,4 @@ Bitcoin ABC version 0.19.5 is now available from:
   <https://download.bitcoinabc.org/0.19.5/>
 
 This release includes the following features and fixes:
-
+ - Deprecated the `-reserveChangeKey` option for `fundrawtransaction` wallet rpc command.